Separated seed into data and functionality
